changed selemtype to long long, main test "stacksize"."top"

+++ stack/stack.h
@@ -1,7 +1,7 @@
 #ifndef _STACH_H_
 #define _STACH_H_
 
-typedef int selemtype;
+typedef long long selemtype;
 
 typedef struct sqstack
 {
